Create useCalendarEvents hook
